requirements.txt
aiohttp==3.7.4
chia_blockchain==1.1.6
nest_asyncio==1.5.1
dateparser==1.0.0
requests==2.25.1
psutil==5.8.0
chia==2.3.1
discord_notify==1.0.0
playsound==1.2.2
prometheus_client==0.10.1
pushover==0.5
PyYAML==5.4.1
telegram_notifier==0.3
